product comment
product comment save -> u can add a comment to a product, if the user has already a comment, it will be deleted and added for the new one. The average score is changed automatically in the addittion with the new comment.

get average score -> new callback to get only the product's average_score